@@ -1,4 +1,4 @@
-declare module "*.png" {
+declare module '*.png' {
   const value: string;
   export default value;
 }
