@@ -1,4 +1,4 @@
-import SR from 'seedrandom';
+import SR from 'seedrandom'
 
 export default abstract class Operation {
   public digit1 = 0;
@@ -9,42 +9,36 @@ export default abstract class Operation {
 
   public abstract color: string;
 
-  public static prng: Function;
+  public static prng = SR(Date.now().toString());
 
-  constructor(level: number) {
-    if (!Operation.prng) {
-      // seed random with date to get real random.
-      // use a random lib with seed to allow later someone to launch
-      // exactly the same sequence than anoter (battle mode)
-      Operation.prng = SR(Date.now().toString());
-    }
+  constructor (level: number) {
     // loop until the generated digits generates
     // an operation that fit the level
     do {
-      this.generateDigits(level);
-    } while (this.digit1 > 99
-      || this.digit2 > 99
-      || this.result > 99
-      || this.stars > level);
+      this.generateDigits(level)
+    } while (this.digit1 > 99 ||
+      this.digit2 > 99 ||
+      this.result > 99 ||
+      this.stars > level)
   }
 
-  static rand(min: number, max: number): number {
-    return Math.floor(min + Operation.prng() * (max - min));
+  static rand (min: number, max: number): number {
+    return Math.floor(min + Operation.prng() * (max - min))
   }
 
   /* eslint-disable class-methods-use-this */
-  public get result(): number {
-    return 0;
+  public get result (): number {
+    return 0
   }
 
-  public toString(): string {
-    return `${this.digit1} ${this.sign} ${this.digit2}`;
+  public toString (): string {
+    return `${this.digit1} ${this.sign} ${this.digit2}`
   }
 
   protected abstract generateDigits(level: number): void;
 
   /* eslint-disable class-methods-use-this */
-  public get stars(): number {
-    return 0;
+  public get stars (): number {
+    return 0
   }
 }
