@@ -1,5 +1,5 @@
 export default interface Score {
-  kind: string,
-  level: number,
-  score: number
+  kind: string;
+  level: number;
+  score: number;
 }
